pyproject.toml
[project]
name = "mlops-lab1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pre-commit>=4.5.1",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.41.0",
]
